STEP 1 추가 구현사항 정리
- final 키워드 추가
- Model 파일 분리
