+++ ParcelInvoiceMaker/ParcelInvoiceMaker/InvoiceViewController.swift
@@ -6,7 +6,7 @@
 
 import UIKit
 
-class InvoiceViewController: UIViewController {
+final class InvoiceViewController: UIViewController {
 
     private let parcelInformation: ParcelInformation
